travis: enable aarch64 builds
While we enable these additional architectures, we limit the code
coverage test to x86_64 and the clang/scan-build check to platforms
which support it.

It is worth noting that the ppc64le and s390x builds were not enabled
at this time due to Travis CI test failures that need to be
investigated.

Signed-off-by: Paul Moore <paul@paul-moore.com>