doc: clarify seccomp_rule_add syscall translation behavior
libseccomp performs a translation step when adding a raw syscall value
to a multi-architecture filter. For instance, when adding __NR_open
(syscall value 2 on x86-64) to a filter containing x86 and x86-64 where
the native ABI is x86-64, the x86 BPF branch will use the value 5
(__NR_open on x86).

This commit adds explicit documentation for the translation step.

Refs https://github.com/seccomp/libseccomp/issues/259.

Signed-off-by: Tudor Brindus <me@tbrindus.ca>