tests: Handle pseudo-syscalls in the sim binary tree test
Now that pseudo-syscalls are being removed from the resultant
BPF filter, we need to test for this in the simulated binary
tree test.  This commit modifies the test to explicitly define
its supported architectures and updates the tests file to
handle PNR syscalls.  Note that the aarch64 architecture does
not define many of the syscalls used in this test, and thus
these syscalls fall through to the default ALLOW action.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>