doc: clarify that syscall must exist in all filter architectures
If a syscall is used in a multi-architecture filter, the syscall must
exist in all the architectures, or -EOPNOTSUPP is returned. For example,
epoll_wait_old has value 215 in x86-64, but does not exist in x86.
Trying to add a filter rule including it in a x86-64/x86 filter will
fail.

This commit clarifies that libseccomp will reject a rule containing such
a case.

Signed-off-by: Tudor Brindus <me@tbrindus.ca>