tests: Explicitly specify the archs in the sim binary tree test
Test 53-sim-binary_tree removes the native architecture and
builds the binary tree on aarch64, ppc64le, and x86_64.  But the
tests file was testing for "all" architectures which led to test
failures on other systems like s390x.  This commit replaces the
"all" arch with only the architectures in the test.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>