tests: use smaller "magic" numbers in 51-live-user_notification
On an x32 based system, and perhaps other 32-bit systems, the magic
value in test 51-live-user_notification was too large and resulted
in a failed comparison even when the userspace notification mechanism
was working properly.

This patch addresses this problem by using the parent process's PID
for the magic number.  For all arches/ABIs we know it is a valid
return value for getpid() while also being one of the few PIDs that
the child process should never legitimately get from a call to
getpid().

This patch also restricts the use of SCMP_SYS() to only the
libseccomp API calls.  This should help us catch arches/ABIs that
don't implement getpid().  I'm also not sure we want to be
responsible for using SCMP_SYS() outside of the libseccomp API.

Acked-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>