tests: limit the number of bpf-sim-fuzz test iterations
The bpf-sim-fuzz test type was always intended to be a fuzzer for the
generated BPF filters but the current implementation fuzzes the
simulator more than the filter itself.  Due to this, let's bump down
the bpf-sim-fuzz iterations from 50 to 5, this is what we currently
do on Travis CI and it seems to be okay.

This change drops the number of tests by 1800 and the runtime from
4m7s to 3m35s on my test system running eight parallel jobs.

Reviewed-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>