tests: add support for the LIBSECCOMP_TSTCFG_JOBS env variable
The LIBSECCOMP_TSTCFG_JOBS environment variable allows the caller to
specify the number of test jobs through the environment as opposed
to the command line.  This is primarily useful for running the tests
through scripts or via the 'make check' command.

The LIBSECCOMP_TSTCFG_JOBS environment variable value behaves the
same as the '-j <JOBS>' command line argument.

Reviewed-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>