doc/man: don't return -1 from main
It is custom for C functions to return a negative integer to denote
an error. This should not be done with main() though, as the return
value is used as an exit code, which is a byte (unsigned char), so
return -1 becomes 255, which is unconventional.

Fix examples where main returns -1 to return 1.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>