doc: fix spacing with .I and punctuation
Use of .I lines causes spaces to be inserted before & after the word.
When words are before or after, that's fine, but when it's punctuation
like parentheses or commas, it looks weird.  Switch to .IR and .RI to
tighten up the display.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Paul Moore <paul@paul-moore.com>

(imported from commit 2cfc5bd29c90a2a9b03ff00ef81b5742a1acf506)