tests: fix 54-live-binary_tree to use binary tree
Apparently, an early implementation of the binary tree optimization
used to enable the feature when the number of rules added was > 16.

The code was later changed to add and use SCMP_FLTATR_CTL_OPTIMIZE,
but the 54-live-binary_tree test case was left as is. So, despite
its name, it is not testing the binary tree.

Fix this, and remove the comment that referred to the old
implementation.

Fixes: 38f04da84748 ("tests: add tests for the binary tree")
Reviewed-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
[PM: restyled the "Fixes" tag]
Signed-off-by: Paul Moore <paul@paul-moore.com>

(imported from commit 5731b3c338f8f18b1d2b3aa300bbcb97af0fb34c)