arch: require arch-syscall-dump from arch-syscall-validate
The arch-syscall-dump is almost always used from arch-syscall-validate
(the only exception I found is when -c and -l are given), so it makes
sense to check that it's available beforehand.

Without this patch, when arch-syscall-dump is not present, the script
produces lots of output with a few errors about the missing
./arch-syscall-dump hidden in the middle of it.

To keep things simple, we require arch-syscall-dump even when it is not
going to be used.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>