python: Link module against shared library
Currently, the libseccomp.a is linked into the python module
statically. This is suboptimal when compared to what dynamic
linking offers. Therefore, make the python module link to
libseccomp.so.

In order to allow test suite find the freshly built library we
need to prepend the LD_LIBRARY_PATH with path to the libtools
libdir (.libs).

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>