Revert "tests: Fix wrong syscall-error in test 29"
Revert commit a707ec98cc2c ("tests: Fix wrong syscall-error in test
29").  The reverted commit properly set the syscall number to -10001 in
a call to seccomp_rule_add_exact(), but due to previously unknown
shortcomings in the test, this led to failures on aarch64 (and possibly
other architectures).  Revert to the previous version of test 29.

Note that this reverted version of test 29 is not functioning as
originally designed, but it is largely benign and the functional test
will pass on all architectures.  (It's generating a BPF filter that will
allow all syscalls on x86.)

Fixes: a707ec98cc2c ("tests: Fix wrong syscall-error in test 29")
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>