tests: fix a test framework race condition and spurious output
Force the creation of the test log and stats files before running the
test, this prevents a race condition where the framework attempts to
output the test results before the test has had a chance to run.

Quiet a grep command in the framework that was resulting in unwanted
output in the test terminal.

Signed-off-by: Paul Moore <paul@paul-moore.com>
(imported from commit 0ab53f95006dcdc525fff175f2eff9f5a5d56883)
Signed-off-by: Paul Moore <paul@paul-moore.com>