python: Rename setup.py to setup.py.in
Rename setup.py to setup.py.in so that the version variables can be
autopopulated by automake/m4.

This will be used in subsequent commits for building python wheels.
Wheels are often built in containers, and the build containers don't
have knowledge of automake, its configurations, and its environment
variables, so we need to pre-populate the version information.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>